Enhance pricing and vendor management in orders
Added properties for pricing and vendor information in
MealService, CreateOrderRequest, and Order classes.
Updated database migrations to reflect these changes.
Introduced OrderCompletedProjection for tracking completed
orders and implemented OrderCompletedEventHandler for
inserting order projections into the database.
These enhancements improve the application's financial
management capabilities for meal orders.

+++ EatUp.Orders/DTO/OrderDTO.cs
@@ -13,7 +13,7 @@ namespace EatUp.Orders.DTO
         public string FoodPackageTitle { get; set; }
         public PaymentStatusEnum PaymentStatus { get; set; }
         public string? PaymentId { get; set; }
-        public float Price { get; internal set; }
+        public double Price { get; internal set; }
         public int Quantity { get; set; }
         public DateTime? CreatedAt { get; set; }
